+++ src/editor-model/commands.ts
@@ -295,6 +295,8 @@ export function move(
 ): boolean {
   options = options ?? { extend: false };
 
+  model.mathfield.adoptStyle = direction === 'backward' ? 'right' : 'left';
+
   if (direction !== 'forward') {
     const [from, to] = getCommandSuggestionRange(model);
     if (from !== undefined && to !== undefined) model.deleteAtoms([from, to]);